requirements.txt
altgraph==0.17
buf==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
contextlib2==0.6.0.post1
cycler==0.10.0
dcs==0.1.1
docopt==0.6.2
future==0.18.2
h5py==2.10.0
idna==2.10
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
numpy==1.19.1
opencv-python==4.4.0.42
panda==0.3.1
pefile==2019.4.18
pgcolorbar==1.0.0
Pillow==7.2.0
psutil==5.7.2
pyaml==20.4.0
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.10
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.13.2
PyQt5-sip==12.8.0
PyQt5-stubs==5.14.2.2
pyqt5-tools==5.13.2.1.6rc1
pyqtgraph==0.11.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.24.0
schema==0.7.2
scipy==1.5.2
six==1.15.0
spline==1.12
tabulate==0.8.2
urllib3==1.25.10
pyinstaller --hidden-import=pkg_resources -w -i ouc.ico -F main.py